Add files via upload
Code opens a file and reads values, goes through various calculations, then displays values based to satisfy the following prompt: 

The James Bufét Water Taxi Company uses a boat on the Caribbean Sea to transport people and goods from the main dock to various islands and ports. The service charge for a trip is $15.00 for each stop and $2.60 per kilometer of travel. Some people are regular customers and granted a VIP card to receive a 25% discount of their trip cost. A GPS records the (x, y) position for each stop. All trips begin at the main dock (0,0). Each leg of the trip is made in a straight line of the shortest distance between stops (including the main dock). A file called “taxi.txt” contains the records for all the trips made in August. Each line in the file contains the information for one customer. The data on each line is formatted as follows: • A binary (0 or 1) value that indicates whether the customer has a VIP card or not. 1 = VIP while 0 = not VIP. • The number of stops, not including returning to the main dock. • Pairs of (x,y) coordinates for each of the stops. 

Your program is to generate a report file summarizing the taxi’s earnings for the month of August. The report must contain the following information on each line: • Trip number
• VIP or not (0 = no, 1 = yes)
• Number of stops (not including return to main dock) • Total distance travelled for the trip
• Total cost charged to the customer
• Cumulative distance travelled since the beginning of the month • Cumulative cost charged to all customers during the month (money the taxi driver collected for the month